Add list_iso method to iso_tools api
