+++ kiwi/iso_tools/iso.py
@@ -16,9 +16,7 @@
 # along with kiwi.  If not, see <http://www.gnu.org/licenses/>
 #
 import os
-import re
 import struct
-import collections
 from tempfile import NamedTemporaryFile
 from collections import namedtuple
 
@@ -28,6 +26,7 @@ from collections import namedtuple
 from builtins import bytes
 
 # project
+from kiwi.iso_tools.cdrtools import IsoToolsCdrTools
 from kiwi.logger import log
 from kiwi.defaults import Defaults
 from kiwi.path import Path
@@ -35,7 +34,6 @@ from kiwi.command import Command
 from kiwi.exceptions import (
     KiwiIsoLoaderError,
     KiwiIsoMetaDataError,
-    KiwiIsoToolError,
     KiwiCommandError
 )
 
@@ -287,8 +285,9 @@ class Iso(object):
         file_count = 0
         offset = 0
         found_id = False
+        iso_tool = IsoToolsCdrTools(self.source_dir)
         with open(isofile, 'rb') as iso:
-            for start in self.isols(isofile):
+            for start in iso_tool.list_iso(isofile):
                 if file_count >= 8:  # check only the first 8 files
                     break
                 file_count += 1
@@ -316,39 +315,6 @@ class Iso(object):
                     'Header ID not found in iso file %s' % isofile
                 )
             return offset * 4
-
-    def isols(self, isofile):
-        """
-        List contents of an ISO image
-
-        :param string isofile: path to the ISO file
-
-        :return: formatted isoinfo result
-        :rtype: dict
-        """
-        listing_type = namedtuple(
-            'listing_type', ['name', 'filetype', 'start']
-        )
-        listing = Command.run(
-            [self._get_isoinfo_tool(), '-R', '-l', '-i', isofile]
-        )
-        listing_result = {}
-        for line in listing.output.split('\n'):
-            iso_entry = re.search(
-                '.*(-[-rwx]{9}).*\s\[\s*(\d+)(\s+\d+)?\]\s+(.*?)\s*$', line
-            )
-            if iso_entry:
-                entry_type = iso_entry.group(1)
-                entry_name = iso_entry.group(4)
-                entry_addr = int(iso_entry.group(2))
-                listing_result[entry_addr] = listing_type(
-                    name=entry_name,
-                    filetype=entry_type,
-                    start=entry_addr
-                )
-        return collections.OrderedDict(
-            sorted(listing_result.items())
-        )
 
     def setup_isolinux_boot_path(self):
         """
@@ -501,22 +467,6 @@ class Iso(object):
                 '%s: no boot catalog found' %
                 iso_header.isofile
             )
-
-    def _get_isoinfo_tool(self):
-        """
-        There are tools by J.Schilling and tools from the community
-        This method searches in all paths which could provide an
-        isoinfo tool. The first match makes the decision
-        """
-        alternative_lookup_paths = ['/usr/lib/genisoimage']
-        isoinfo = Path.which('isoinfo', alternative_lookup_paths)
-        if isoinfo:
-            return isoinfo
-
-        raise KiwiIsoToolError(
-            'No isoinfo tool found, searched in PATH: %s and %s' %
-            (os.environ.get('PATH'), alternative_lookup_paths)
-        )
 
     @staticmethod
     def _read_iso_sector(sector, handle):
